Better error messages, dependency installation in buildall.sh.

+++ buildall.sh
@@ -1,16 +1,39 @@
 #!/bin/bash
-runghc Setup.hs configure --user && runghc Setup.hs build && runghc Setup.hs install
 
+runghc Setup.hs configure --user
+
+if [ "$?" != "0" ] ; then
+    echo "Can't configure Haste; you seem to be missing a dependency or two."
+    echo "Attempting to get them from Hackage..."
+    cabal update
+    cabal install mtl cereal containers bytestring filepath directory array ghc-paths
+
+    runghc Setup.hs configure --user
+    if [ "$?" != "0" ] ; then
+        echo "Still can't configure Haste, and I have no idea why. :("
+        exit 1
+    fi
+fi
+
+runghc Setup.hs build && runghc Setup.hs install
+if [ "$?" != "0" ] ; then
+    echo "Unable to build Haste, and I have no idea why. :("
+    exit 1
+fi
+
+# This step can't fail. I hope. At least not when downloading libs. Probably.
 if [ "$1" != "" ] ; then
-  ./buildlibs.sh $1
+    echo "Local lib build requested, using GHC source from $1"
+    ./buildlibs.sh $1
 else
-  TMPDIR=$(mktemp -d)
-  wget -O$TMPDIR/haste-libs.tar.bz2 http://ekblad.cc/haste-libs.tar.bz2
-  pushd .
-  cd
-  tar -xjf $TMPDIR/haste-libs.tar.bz2
-  popd
-  rm -r $TMPDIR
+    echo "Downloading base libraries from ekblad.cc"
+    TMPDIR=$(mktemp -d)
+    wget -O$TMPDIR/haste-libs.tar.bz2 http://ekblad.cc/haste-libs.tar.bz2
+    pushd .
+    cd
+    tar -xjf $TMPDIR/haste-libs.tar.bz2
+    popd
+    rm -r $TMPDIR
 fi
 
 pushd .
